refactor: Restructure the _auth routes and related code
