+++ app/routes/_auth.logout/route.tsx
