+++ app/routes/_auth.register/userSchema.ts
@@ -1,3 +1,4 @@
+import { withZod } from "@remix-validated-form/with-zod";
 import { z } from "zod";
 
 /**
@@ -11,14 +12,6 @@ export const userSchemaRegister = z.object({
     .max(24, "Must contain 24 chars max"),
 });
 
-/**
- * User Schema used in Login Form
- */
-export const userSchemaLogin = z.object({
-  email: z.string().email(),
-  password: z.string().min(1, "Must provide password"),
-});
-
 export type RegisterAccountAuth = z.infer<typeof userSchemaRegister>;
 
-export type LoginAuth = z.infer<typeof userSchemaLogin>;
+export const validator = withZod(userSchemaRegister);
